move serloDomain to helpers
